python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
psycopg2-binary==2.9.9
requests==2.32.3
beautifulsoup4==4.12.3
pytest==8.3.3
httpx==0.27.2
